feat!: allow use of custom Intl.Segmenter implementation
BREAKING CHANGE: The Tally constructor now takes a single options object instead of a locale string. The `locale` option has also been renamed to `locales` to indicate that multiple locales can be provided. This matches the behavior of `Intl.Segmenter`. To set a locale, pass an object like `{locales: 'en'}` to the Tally constructor.